% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_AIC.R
\name{cp_AIC}
\alias{cp_AIC}
\title{Compute Akaike's information criterion}
\usage{
cp_AIC(ll, k)
}
\arguments{
\item{ll}{log-likelihood of the model of interest}

\item{k}{number of estimated parameters by the model}
}
\value{
AIC stored as a numeric unit vector of length 1.
}
\description{
Computes Akaike's information criterion for comparing competing models.
}
\examples{
# Fit some model
lm_out <- lm(mpg ~ cyl + disp, data = mtcars)

# Compute AIC with your function
AIC_M <- cp_AIC(
    ll = logLik(lm_out),
    k = length(coef(lm_out)) + 1 # intercept + reg coefs + error variance
)
}
\author{
Edoardo Costantini, 2023
}
