% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_pc_scores.R
\name{cp_pc_scores}
\alias{cp_pc_scores}
\title{Compute principal component scores}
\usage{
cp_pc_scores(X_train, X_valid, Q)
}
\arguments{
\item{X_train}{Matrix of IV values in the training dataset.}

\item{X_valid}{Matrix of IV values in the validation dataset.}

\item{Q}{atomic numeric vector containing the number of principal components to be used}
}
\value{
Description of function output
}
\description{
Given a training and validation data set, it estimates the component loadings on the training data and returns the component scores on both the training and validation data.
}
\details{
This function does such and such.
}
\author{
Edoardo Costantini, 2023
}
