% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_thrs_LLS.R
\name{cp_thrs_LLS}
\alias{cp_thrs_LLS}
\title{Compute threshold values based on Log-likelihood values}
\usage{
cp_thrs_LLS(dv, ivs, fam)
}
\arguments{
\item{dv}{Vector of dependent variable values}

\item{ivs}{Matrix of predictor values}

\item{fam}{GLM framework for the dv}
}
\value{
A vector of bivariate association measures between \code{dv} and \code{ivs}.
}
\description{
Produces a vector of threshold values that define active predictors.
}
\details{
This function does such and such.
}
\examples{
# Example inputs
dv <- mtcars[, 1]
ivs <- mtcars[, -1]

}
\author{
Edoardo Costantini, 2023
}
