% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_thrs_NOR.R
\name{cp_thrs_NOR}
\alias{cp_thrs_NOR}
\title{Compute normalized association measures}
\usage{
cp_thrs_NOR(dv, ivs, s0_perc = NULL)
}
\arguments{
\item{dv}{Vector of dependent variable values}

\item{ivs}{Matrix of predictor values}

\item{s0_perc}{Factor for the denominator of association statistic, between 0 and 1: the percentile of standard deviation values added to the denominator. Default is 0.5 (the median)}
}
\value{
A vector of bivariate association measures between \code{dv} and \code{ivs}.
}
\description{
A function to compute the normalized bivariate association measures between a \code{dv} and a collection of \code{ivs}.
}
\details{
This function is based on the function \code{cor.func} in the package \code{superpc}.
}
\examples{
# Example inputs
dv <- mtcars[, 1]
ivs <- mtcars[, -1]
s0_perc <- 0

}
\references{
Bair E, Hastie T, Paul D, Tibshirani R (2006). “Prediction by supervised principal components.” J. Am. Stat. Assoc., 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
