% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_validation_fit.R
\name{cp_validation_fit}
\alias{cp_validation_fit}
\title{Compute fit measure(s) on the validation data set}
\usage{
cp_validation_fit(y_train, y_valid, X_train, X_valid, fam, fit_measure)
}
\arguments{
\item{y_train}{Vector of DV values in the training dataset.}

\item{y_valid}{Vector of DV values in the validation dataset.}

\item{X_train}{Matrix of IV values in the training dataset. Can also be set to 1 to obtain the log-likelihood of the new data under the null model.}

\item{X_valid}{Matrix of IV values in the validation dataset. If \code{X_train} is set to 1 to obtain the log-likelihood of the new data under the null model, \code{X_valid} is ignored.}

\item{fam}{GLM framework for the dv.}

\item{fit_measure}{Character vector indicating which fit measure should be computed (F, LRT, AIC, BIC, PR2 ,MSE)}
}
\value{
Description of function output
}
\description{
Given a training and validation data set, it computes a target fit measure on the validation data set.
}
\details{
The validation data set can be specified to be the same as the training data set if desired.
}
\author{
Edoardo Costantini, 2023
}
