% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.average.scores.R
\name{cv.average.scores}
\alias{cv.average.scores}
\title{Averaging results from CV}
\usage{
cv.average.scores(cv_array, test)
}
\arguments{
\item{cv_array}{A an array of q nt k dimensionality.}

\item{test}{The type of test stored in the array}
}
\value{
A list of three q nt matrices.
\itemize{
\item scor: contains the average CV scores across the K folds
\item scor.upr: contains the average CV scores across the K folds + 1 standard deviation
\item scor.lwr: contains the average CV scores across the K folds - 1 standard deviation
}
}
\description{
A low-level function to average results from an array of CV results.
}
\details{
The input of this function is an array of q nt k, where q is the number of principal components, nt is the number of thresholds and k is the number of folds.
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
