% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.choice.R
\name{cv.choice}
\alias{cv.choice}
\title{Cross-validation choice}
\usage{
cv.choice(scor, scor.lwr, scor.upr, K, test)
}
\arguments{
\item{scor}{A description of the first argument}

\item{scor.lwr}{A description of the second argument}

\item{scor.upr}{A description}

\item{K}{A description}

\item{Test}{A description}
}
\value{
A list of two unit vectors:
\itemize{
\item default = The default choice,
\item oneSE = The choice based on the one standard error rule.
}
}
\description{
Extracting the CV choices of SPCR parameters.
}
\details{
Given a matrix of npcs * thrsh, returns the best choice based on the type of test (best overall and 1se rule versions)
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
