% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gspcr.R
\name{cv.gspcr}
\alias{cv.gspcr}
\title{Cross-validation of Generalized Principal Component Regression}
\usage{
cv.gspcr(
  dv,
  ivs,
  fam = "gaussian",
  thrs = c("LLS", "pseudoR2", "normalized")[1],
  nthrs = 10,
  maxnpcs = 3,
  K = 5,
  test = c("LRT", "F", "MSE")[2],
  max.features = ncol(ivs),
  min.features = 5,
  oneSE = TRUE
)
}
\arguments{
\item{dv}{Vector of dependent variable values}

\item{ivs}{Matrix of predictor values}

\item{fam}{GLM framework for the dv}

\item{thrs}{Type of threshold to be used}

\item{nthrs}{Number of threshold values to be used}

\item{maxnpcs}{= Maximum number of principal components to be used}

\item{K}{Number of folds for the K-fold cross-validation procedure}

\item{test}{Type of measure to cross-validate}

\item{max.features}{Maximum number of features that can be selected}

\item{min.features}{Minimum number of features that can be selected}

\item{oneSE}{Whether the results with the 1SE rule should be stored}
}
\value{
Returns an object of class gspcr.
}
\description{
Use K-fold cross-validation to decide on the number of principal components and the threshold value for GSPCR.
}
\details{
This function does such and such.
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
