% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_average.R
\name{cv_average}
\alias{cv_average}
\title{Average fit measures computed in the K-fold cross-validation procedure}
\usage{
cv_average(cv_array, fit_measure)
}
\arguments{
\item{cv_array}{A an array of q nt k dimensionality.}

\item{fit_measure}{The type of fit measure stored in the array}
}
\value{
A list of three q nt matrices.
\itemize{
\item scor: contains the average CV scores across the K folds
\item scor_upr: contains the average CV scores across the K folds + 1 standard deviation
\item scor_lwr: contains the average CV scores across the K folds - 1 standard deviation
}
}
\description{
A low-level function to average results from an array of K-fold CV fit measures.
}
\details{
The input of this function is an array of q nt k, where q is the number of principal components, nt is the number of thresholds and k is the number of folds.
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
