% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_choose.R
\name{cv_choose}
\alias{cv_choose}
\title{Cross-validation choice}
\usage{
cv_choose(scor, scor.lwr, scor.upr, K, test)
}
\arguments{
\item{scor}{A \eqn{npcs \times nthrs} matrix of K-fold CV scores.}

\item{scor.lwr}{A \eqn{npcs \times nthrs} matrix of score lower bounds.}

\item{scor.upr}{A \eqn{npcs \times nthrs} matrix of score upper bounds.}

\item{K}{The number of folds used for K-fold cross-validation.}

\item{test}{The type of score to compute for the cross-validation procedure.}
}
\value{
A list of two unit vectors:
\itemize{
\item default = The default choice.
\item oneSE = The choice based on the one standard error rule.
}
}
\description{
Extracting the CV choices of SPCR parameters.
}
\details{
Given a matrix of \eqn{npcs \times nthrs}, returns the best choice based on the type of test (best overall and 1se rule versions)
}
\author{
Edoardo Costantini, 2023
}
