% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_gspcr.R
\name{cv_gspcr}
\alias{cv_gspcr}
\title{Cross-validation of Generalized Principal Component Regression}
\usage{
cv_gspcr(
  dv,
  ivs,
  fam = "gaussian",
  thrs = c("LLS", "PR2", "normalized")[1],
  nthrs = 10,
  npcs_range = 1:3,
  K = 5,
  fit_measure = c("F", "LRT", "AIC", "BIC", "PR2", "MSE")[1],
  max_features = ncol(ivs),
  min_features = 5,
  pca_implementation = "PCAmixdata",
  oneSE = TRUE
)
}
\arguments{
\item{dv}{Vector of dependent variable values}

\item{ivs}{Matrix of predictor values}

\item{fam}{GLM framework for the dv}

\item{thrs}{Type of threshold to be used}

\item{nthrs}{Number of threshold values to be used}

\item{npcs_range}{range of number of principal components to be used}

\item{K}{Number of folds for the K-fold cross-validation procedure}

\item{fit_measure}{Type of measure to cross-validate.}

\item{max_features}{Maximum number of features that can be selected}

\item{min_features}{Minimum number of features that can be selected}

\item{pca_implementation}{whether \code{PCAmixdata} or an \code{independent} implementation of PCAmix should be used to compute PCA.}

\item{oneSE}{Whether the results with the 1SE rule should be stored}
}
\value{
Returns an object of class \code{gspcr}.
}
\description{
Use K-fold cross-validation to decide on the number of principal components and the threshold value for GSPCR.
}
\details{
The possible threshold types are:
\itemize{
\item \code{LLS}
\item \code{PR2} - The Cox and Snell generalized R-squared is computed for the GLMs between \code{dv} and every column in \code{ivs}. Then, the square root of these values is used as to obtain the threshold values. For more information about the computation of the Cox and Snell R2 see the help file for \code{\link[=cp_gR2]{cp_gR2()}}. When using this measure for simple linear regressions (with continuous \code{dv} and \code{ivs}) is equivalent to the regular R-squared. Therefore, it can be thought of as equivalent to the bivariate correlations between \code{dv} and \code{ivs}.
\item \code{normalized}
}
The possible fit measures are:
\itemize{
\item \code{F}
\item \code{LRT}
\item \code{AIC}
\item \code{BIC}
\item \code{PR2}
\item \code{MSE}
}
}
\examples{
# Example input values
dv <- mtcars[, 1]
ivs <- mtcars[, -1]
thrs <- "PR2"
nthrs <- 5
fam <- "gaussian"
npcs_range <- 1:3
K <- 3
fit_measure <- "F"
max_features <- ncol(ivs)
min_features <- 1
oneSE <- TRUE
pca_implementation = "PCAmixdata"

# Example usage
out_cont <- cv_gspcr(
   dv = GSPCRexdata$y$cont,
   ivs = GSPCRexdata$X$cont,
   fam = "gaussian",
   nthrs = 5,
   npcs_range = 1:3,
   K = 3,
   fit_measure = "F",
   thrs = "normalized",
   min_features = 1,
   max_features = ncol(GSPCRexdata$X$cont),
   pca_implementation = "PCAmixdata",
   oneSE = TRUE
)

}
\references{
Bair, E., Hastie, T., Paul, D., & Tibshirani, R. (2006). Prediction by supervised principal components. Journal of the American Statistical Association, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
