% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_univ_mods.R
\name{est_univ_mods}
\alias{est_univ_mods}
\title{Estimate univariate standard GLMs.}
\usage{
est_univ_mods(dv, ivs, fam)
}
\arguments{
\item{dv}{Vector of dependent variable values}

\item{ivs}{Matrix of predictor values}

\item{fam}{GLM framework for the dv}
}
\value{
A list of:
\itemize{
\item \code{ll0}: log-likelihoods for the null model
\item \code{lls}: log-likelihoods for all the univariate models
\item \code{coefs}: if \code{dv} and \code{ivs} are continuous, standardized univariate regression coefficients
}
}
\description{
Given a dependent variable, a set of possible predictors and a GLM family, this function estimates a null GLM and all of the univariate GLMs.
}
\details{
This function does such and such.
}
\author{
Edoardo Costantini, 2023
}
