% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{gpca}
\alias{gpca}
\title{Generalized Principal Component Analysis}
\usage{
gpca(X_tr, X_new, npcs = 1)
}
\arguments{
\item{argument_1}{A description of the first argument}

\item{argument_2}{A description of the second argument}
}
\value{
Description of function output
}
\description{
A high level description of the function.
}
\details{
This function does such and such.
}
\examples{
# Example inputs
data(wine, package = "FactoMineR")
X <- wine[, c(1, 2, 16, 22)]
X$Label <- factor(X$Label)
X$Soil <- factor(X$Soil)
X_tr <- X[1:15, ]
X_new <- X[16:21, ]
npcs <- 2

# Example use
gpca(
    X_tr = X[1:15, ],
    X_new = X[16:21, ],
    npcs = 2
)
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
