% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_baseline.R
\name{LL_baseline}
\alias{LL_baseline}
\title{Baseline category logistic regression log-likelihood}
\usage{
LL_baseline(y, syst_comp)
}
\arguments{
\item{y}{A multi-categorical factor or its disjunctive table representation.}

\item{syst_comp}{A matrix storing all of the systematic components for a baseline category logistic regression.}
}
\value{
Atomic vector of length 1 containing the log-likelihood value.
}
\description{
Computes the baseline category logistic regression log-likelihood given a nominal categorical variable and the corresponding GLM linear predictor values.
}
\details{
A disjunctive table is a matrix representation of a multi-categorical variable. The dimensionality of the matrix is i times j, with i = number of observations, and j = number of categories. \code{y_{ij}} is equal to 1 if observation i responded with category j, and it is equal to 0 otherwise.
}
\author{
Edoardo Costantini, 2023
}
