% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_binomial.R
\name{LL_binomial}
\alias{LL_binomial}
\title{Binomial log-likelihood}
\usage{
LL_binomial(y, lgt)
}
\arguments{
\item{y}{Vector (or factor) of values on a binary dependent variable.}

\item{lgt}{Vector of GLM linear predictor values.}
}
\value{
Atomic vector of length 1 containing the log-likelihood value.
}
\description{
Computes the binomial log-likelihood given a response vector and corresponding GLM linear predictor values.
}
\details{
The log-likelihood equation is based on Agresti (2002, p. 192).
}
\references{
Agresti, A. (2012). Categorical data analysis (Vol. 792). John Wiley & Sons.
}
\author{
Edoardo Costantini, 2022
}
