% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_cumulative.R
\name{LL_cumulative}
\alias{LL_cumulative}
\title{Cumulative link model log-likelihood}
\usage{
LL_cumulative(y, syst_comp)
}
\arguments{
\item{y}{Vector (or factor) of values on a binary dependent variable.}

\item{syst_comp}{Matrix of GLM systematic component values.}
}
\value{
Atomic vector of length 1 containing the log-likelihood value.
}
\description{
Computes the log-likelihood given an ordered categories response vector and corresponding GLM linear predictor values.
}
\details{
The log-likelihood equation is based on Agresti (2002, p. 192).
}
\references{
Agresti, A. (2012). Categorical data analysis (Vol. 792). John Wiley & Sons.
}
\author{
Edoardo Costantini, 2022
}
