% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_gaussian.R
\name{LL_gaussian}
\alias{LL_gaussian}
\title{Normal log-likelihood}
\usage{
LL_gaussian(y, y_hat, mod)
}
\arguments{
\item{r}{Vector of residuals}

\item{s}{Residual standard deviation}
}
\value{
The log-likelihood value
}
\description{
Computes the normal log-likelihood
}
\details{
Given
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2022
}
