% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_poisson.R
\name{LL_poisson}
\alias{LL_poisson}
\title{Poisson regression log-likelihood}
\usage{
LL_poisson(y, syst_comp)
}
\arguments{
\item{y}{Vector of values on a count dependent variable.}

\item{syst_comp}{Vector of GLM linear predictor values.}
}
\value{
Atomic vector of length 1 containing the log-likelihood value.
}
\description{
Computes the Poisson regression log-likelihood of a vector of observed values given the GLM systematic component.
}
\references{
Agresti, A. (2012). Categorical data analysis (Vol. 792). John Wiley & Sons.
}
\author{
Edoardo Costantini, 2023
}
