% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{loglike_norm}
\alias{loglike_norm}
\title{Normal log-likelihood}
\usage{
loglike_norm(r, s)
}
\arguments{
\item{r}{Vector of residuals}

\item{s}{Residual standard deviation}
}
\value{
A matrix of the infile
}
\description{
Computes the normal log-likelihood
}
\details{
Given
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2022
}
