% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gspcr.R
\name{plot.gspcr}
\alias{plot.gspcr}
\title{Plot the gspcr solution path}
\usage{
\method{plot}{gspcr}(
  gspcr.out,
  labels = TRUE,
  errorBars = FALSE,
  discretize = TRUE,
  print = TRUE
)
}
\arguments{
\item{gspcr.out}{The output of a cv.gspcr() call.}

\item{labels}{Logical value. FALSE hides the labels of the points indicating the number of components used. The default is TRUE.}

\item{errorBars}{Logical value. TRUE shows the errorBars for each point. The default is FALSE.}

\item{discretize}{Logical value. TRUE treats the X-axis as a discrete measure that facilitates comparing solution paths between different test types. The default is TRUE.}

\item{print}{Logical value. TRUE prints the plot when the function is called. The default is TRUE.}
}
\value{
A scatterplot of ggplot class
}
\description{
Produces a scatterplot showing the CV score obtained by cv.gspcr (Y-axis) with different threshold values (X-axis) for a different number of components (lines).
}
\details{
The bounds defining the error bars are computed by cv.gspcr(). First, the K-fold cross-validation score of the statistic of interest (e.g., the F score, the MSE) is computed. Then, the standard deviation of the statistic across the K folds is computed. Finally, the bounds used for the error bars are computed by summing and subtracting this standard deviation to and from the K-fold cross-validation score of the statistic.
}
\author{
Edoardo Costantini, 2023
}
