% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gspcrout.R
\name{predict.gspcrout}
\alias{predict.gspcrout}
\title{Predict GSPCR model dependent variable scores}
\usage{
\method{predict}{gspcrout}(object, ..., newdata = NULL)
}
\arguments{
\item{object}{An object of class \code{gspcr}.}
}
\value{
Description of function output
}
\description{
Predicts dependent variable values based on (new) predictor variables values.
}
\details{
This function does such and such.
}
\author{
Edoardo Costantini, 2023
}
