% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_AIC.R
\name{cp_AIC}
\alias{cp_AIC}
\title{Compute Akaike's information criterion}
\usage{
cp_AIC(ll, k)
}
\arguments{
\item{ll}{log-likelihood of the model of interest}

\item{k}{number of estimated parameters by the model}
}
\value{
AIC stored as a numeric unit vector of length 1.
}
\description{
Computes Akaike's information criterion for comparing competing models.
}
\author{
Edoardo Costantini, 2023
}
