% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_pc_scores.R
\name{cp_pc_scores}
\alias{cp_pc_scores}
\title{Compute principal component scores}
\usage{
cp_pc_scores(X_train, X_valid, Q)
}
\arguments{
\item{X_train}{data.frame of IV values in the training dataset}

\item{X_valid}{data.frame of IV values in the validation dataset}

\item{Q}{numeric vector of length 1 storing the number of principal components to be used}
}
\value{
list of two objects:
\itemize{
\item \code{PC_tr} a n times Q matrix of PC scores for the observations in the training data
\item \code{PC_va} a n times Q matrix of PC scores for the observations in the validation data
}
}
\description{
Given a training and validation data set, it estimates the component loadings on the training data and returns the component scores on both the training and validation data.
}
\details{
The function can be used to estimate the PC loadings on training data and project the validation data on the same principal component axis. It can also be used to compute PC scores on a single data set by providing it as both the training and validation data.
}
\author{
Edoardo Costantini, 2023
}
