% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_gspcr.R
\name{cv_gspcr}
\alias{cv_gspcr}
\title{Cross-validation of Generalized Principal Component Regression}
\usage{
cv_gspcr(
  dv,
  ivs,
  fam = "gaussian",
  thrs = c("LLS", "PR2", "normalized")[1],
  nthrs = 10,
  maxnpcs = 3,
  K = 5,
  fit_measure = c("F", "LRT", "AIC", "BIC", "PR2", "MSE")[1],
  max_features = ncol(ivs),
  min_features = 5,
  oneSE = TRUE
)
}
\arguments{
\item{dv}{Vector of dependent variable values}

\item{ivs}{Matrix of predictor values}

\item{fam}{GLM framework for the dv}

\item{thrs}{Type of threshold to be used}

\item{nthrs}{Number of threshold values to be used}

\item{maxnpcs}{Maximum number of principal components to be used}

\item{K}{Number of folds for the K-fold cross-validation procedure}

\item{fit_measure}{Type of measure to cross-validate.}

\item{max_features}{Maximum number of features that can be selected}

\item{min_features}{Minimum number of features that can be selected}

\item{oneSE}{Whether the results with the 1SE rule should be stored}
}
\value{
Returns an object of class \code{gspcr}.
}
\description{
Use K-fold cross-validation to decide on the number of principal components and the threshold value for GSPCR.
}
\details{
The possible threshold types are:
\itemize{
\item \code{LLS}
\item \code{PR2}
\item \code{normalized}
}
The possible fit measures are:
\itemize{
\item \code{F}
\item \code{LRT}
\item \code{AIC}
\item \code{BIC}
\item \code{PR2}
\item \code{MSE}
}
}
\examples{
# Example input values
dv <- mtcars[, 1]
ivs <- mtcars[, -1]
thrs <- "PR2"
nthrs <- 5
fam <- "gaussian"
maxnpcs <- 10
K <- 3
fit_measure <- "F"
max_features <- ncol(ivs)
min_features <- 1
oneSE <- TRUE

# Example usage
out_cont <- cv_gspcr(
   dv = GSPCRexdata$y$cont,
   ivs = GSPCRexdata$X,
   fam = "gaussian",
   nthrs = 5,
   maxnpcs = 5,
   K = 3,
   fit_measure = "F",
   thrs = "normalized",
   min_features = 1,
   max_features = ncol(GSPCRexdata$X),
   oneSE = TRUE
)

}
\references{
Bair, E., Hastie, T., Paul, D., & Tibshirani, R. (2006). Prediction by supervised principal components. Journal of the American Statistical Association, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
