% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{gpca}
\alias{gpca}
\title{Generalized Principal Component Analysis}
\usage{
gpca(X_tr, npcs = 1, scale = "standard")
}
\arguments{
\item{X_tr}{data.frame of training data}

\item{npcs}{number of principal components to keep}

\item{scale}{either "standard" or "MLE"}
}
\value{
Description of function output
}
\description{
A high level description of the function.
}
\details{
This function does such and such.
}
\examples{
# Example inputs
X_tr <- iris
npcs <- 2
scale <- "standard"

# Example use
gpca(
    X_tr = iris,
    npcs = 2
)
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2023
}
