% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspcr-package.R
\docType{package}
\name{gspcr-package}
\alias{gspcr}
\alias{gspcr-package}
\title{gspcr: Generalized Supervised Principal Component Regression}
\description{
The sparse principal component regression is computed. The regularization parameters and number of components are optimized by cross-validation.
}
\author{
\strong{Maintainer}: Edoardo Costantini \email{costantini.edoardo@yahoo.com} (\href{https://orcid.org/0000-0001-9581-9913}{ORCID})

}
\keyword{internal}
