% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_binomial.R
\name{LL_binomial}
\alias{LL_binomial}
\title{Normal log-likelihood}
\usage{
LL_binomial(y, lgt)
}
\arguments{
\item{r}{Vector of residuals}

\item{s}{Residual standard deviation}
}
\value{
The log-likelihood value
}
\description{
Computes the binomial log-likelihood given a vector of probabilities
}
\details{
Given
}
\references{
Such, S. (2006). Such and such. Journal such and such, 101(473), 119-137.
}
\author{
Edoardo Costantini, 2022
}
