% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gspcrout.R
\name{predict.gspcrout}
\alias{predict.gspcrout}
\title{Predict GSPCR model dependent variable scores}
\usage{
\method{predict}{gspcrout}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{gspcr}.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Vector of prediction in "response" format for numerical data and probability of class membership for categorical data
}
\description{
Predicts dependent variable values based on (new) predictor variables values.
}
\author{
Edoardo Costantini, 2023
}
